ASICFWDG-1778 feature: Add cray_diags_fabric.rpm to UAN image build
Description:
Due to changes in Shasta's install process for V1.4, the slingshot fabric diagnostic, dgnettest, needed to change it's install method due to the old Ansible method going away. Now, the install process for the cray_diags_fabric rpm will be to include this file in the UAN image and to use slurm's broadcast feature to deploy the test out to nodes for testing.
Documentation:
N/A

Testing:
Code was tested to build in the DST pipeline, and the image was verified out on Gamora to have the addtional files within.