CASMCMS-6704:
The initialization of three variables was moved out of a block.
The variables might not be correctly initialized inside the block if
the when condition is not satsified. Moving outside of the block
allows them to be initialized properly.

The SLS has an entry 'uan01-can' to which it tacks on '.can' to form
'uan01-can.can'. This play was looking for 'uan01-can.' and not
'uan01-can.can'. It nows looks for 'uan01-can.can'.

This was one potential solution. The other was to
change the SLS entry to be uan01. Then the DNS entry will be
'uan01.can' and will match what the play is looking for.

The play nows ignores the 'device not running' error it receives after
ifup'ing vlan007. The error returns a 158 error code. The vlan007
interface comes up. This is a 'hack' allowing us to move past this
error. It may not be the right decision.