CASMCMS-6704: Fix the UAN play
Check SLS for the CAN element, using the correct index which is 2 not 0.
Previously, this play was not getting the CAN information.

Remove the can-up and can-down calls as they are not needed when using
ifroute to set the CAN route.

Create a bond0 ethernet device.

For ifroute-vlan007 remove the 'table 3' as it is not used in Shasta 1.4
for routing. Remove the netmask.