CASMUSER-2926: Pin CSM versions to use CSM 1.0 rpms
Prevent a traceback error when using CSM 1.2 rpms on a CSM 1.0 system.
Due to changes in the interface in that microservice, the CSM 1.0 versions
of these rpms need to be used when the shasta is running CSM 1.0.

This fix is necessary to support the UKMet system with SS11. This workaround
should be removed when CASMTRIAGE-2859 is addressed. The fix for that issue
may involve CSM providing their own CFS layer for their roles.