CASMUSER-2932: Import the HPE signing key and add CFS tests
Instead of importing a CSM role that was missing from the UAN CFS image,
replicate the tasks to read and import the HPE signing key for rpms.

Add a test layer that will simulate CFS for both CFS image customization
and image personalization as well as run basic syntax checks. The tests
run with the ansible-playbook --check mode. Not all of the ansible roles
support --check mode so future work is required to make sure adequate
test coverage is achieved.

site.yml was reorganized for clarity and to make sure roles such as
motd and shadow only run during CFS image personalization.