CASMUSER-3021 fix mountain NMN failure and CHN wicked logic
There is code for setting up the NMN Gateway for Mountain
cabinets that does not appear to be used and breaks boots on
non-mountain cabinets.  This should be reviewed to see if it
should just be taken out entirely, but for now, it needs a 'when'
clause to tell it not to run if the required fact is not defined.

There is also wicked logic for CHN that only allows wicked to be
used for CHN if there are ifcfg-xxx files for HSN.  The code to
set up the registered result for this is never reached on CAN
systems, resulting in a boot time failure of the ansible.  The
check is now inverted so it looks for CHN first, then whether
ifcfg-xxx file exists.